+++ CHANGELOG.md
@@ -7,7 +7,8 @@ Note that since we don't clearly distinguish between a public and private interf
 ## [Unreleased]
 
 - Add parameter for to display aromatic bonds as dashes next to solid cylinder/line.
-
+- Add backbone representation
+- Fix outline in orthographic mode and set default scale to 2.
 
 ## [v2.0.7] - 2021-06-23